feat: implement and relocate event guide components
- Move EventHero, ThemeSpotlight, CuratedGear, and RelatedEvents to src/features/events/components/
- Refactor styling to use CSS utility classes for complex gradients and overlays
- Ensure compliance with UI audit rules by removing hardcoded HEX values and inline styles
- Update imports in EventGuide and ComponentPreview
- Update visual regression snapshots for research and event-guide pages
